requirements.txt
# Suna-Lite
openai>=1.0.0
click>=8.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.15.0
pyyaml>=6.0.0
rich>=13.0.0
python-dotenv>=1.0.0
lxml>=4.9.0
html5lib>=1.1
aiohttp>=3.8.0
asyncio-throttle>=1.0.0
tavily-python>=0.5.0